refator(home): modify to fit requirements
- remove calendar and dropdowns for time
- add timestamp input field
- remove unnecessary methods relating to date and time
- modify data to be sent during POST
- generate timestamp on the server side
- update tests
- change timestamp data type to bigInteger

+++ migrations/20161007174100_keyvalue.js
@@ -3,7 +3,7 @@ exports.up = function( knex ) {
 	return knex.schema.createTable( 'Entries', function ( table ) {
 		table.string( 'key' );
 		table.string( 'value' );
-		table.integer( 'timestamp' ); // integer instead of timestamp so that it will accept unix values
+		table.bigInteger( 'timestamp' ); // integer instead of timestamp so that it will accept unix values
 	} );
 };
